+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ## [Unreleased]
 
+
+
+## [0.2.1] - 2016-04-07
+
 ### Added
 
 - assigning `nil` via `Tree#[]=` removes a child tree, similar to the assignment
@@ -34,5 +38,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 Initial release
 
 
-[Unreleased]: https://github.com/marcelotto/sycamore/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/marcelotto/sycamore/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/marcelotto/sycamore/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/marcelotto/sycamore/compare/v0.1.0...v0.2.0